# Reduce per-row allocations