Modify several internal functions to avoid allocating unnecessary memory. This commit employs three strategies:

- Allocate a scans slice only when scanning the first row, and store it inside the RowScanner to reuse on subsequent rows.
- When scanning into slices, use more sophistocated reflection code to extend the slie while avoid temporary slice header allocations.
- When scanning into slices of non-pointers, scan directly into the slice index, rather than allocate a new value and copy after scanning.

-----

I used a simple but (hopefully) vaguely representative benchmarking program to measure the impact. 

Hyperfine shows an end-to-end runtime reduction of 40%.
```

# Without change
$ go build main.go && hyperfine --warmup=1 "./main -bench"           12:25:17
Benchmark 1: ./main -bench
  Time (mean ± σ):     15.984 s ±  0.115 s    [User: 93.516 s, System: 20.190 s]
  Range (min … max):   15.825 s … 16.258 s    10 runs
 
# With change
$ go build main.go && hyperfine --warmup=1 "./main -bench"              12:28:54
Benchmark 1: ./main -bench
  Time (mean ± σ):      9.186 s ±  0.094 s    [User: 60.400 s, System: 8.723 s]
  Range (min … max):    9.020 s …  9.385 s    10 runs
```

This also reduced allocated objects (when making 1000 queries returning 1000 rows each) by roughly 3 million and allocated space by ~150MB. (I can share the pprof SVGs if you want.)

Here's the (condensed) text: of the benchmark, if you want to replicate.
```golang
type Person struct {
	Name string
	Age  sql.Null[int64]
}

type Record struct {
	ID int64
	Person
	CreatedAt time.Time
}

func main() {
	ctx := context.Background()

	numThreads := 10
	numIters := 5000
	var wg sync.WaitGroup
	wg.Add(numThreads)
	for t := 0; t < numThreads; t++ {
		go func(t int) {
			_, err := runQueries(ctx, numIters)
			if err != nil {
				fmt.Fprintf(os.Stderr, "QueryRow failed: %v\n", err)
				os.Exit(1)
			}
			wg.Done()
		}(t)
	}
	wg.Wait()
}

func runQueries(ctx context.Context, numIters int) ([]Record, error) {
	conn, err := pgx.Connect(ctx, "postgres://zolstein:password@localhost:5432/db")
	if err != nil {
		fmt.Fprintf(os.Stderr, "Unable to connect to database: %v\n", err)
		os.Exit(1)
	}
	defer conn.Close(ctx)

	records := make([]Record, 1000)
	for i := 0; i < numIters; i++ {
		records = records[:0]
		rows, err := conn.Query(ctx, "select id, name, age, created_at from people") // Returns 1000 rows
		_ = rows
		if err != nil {
			return nil, err
		}
		err = pgxscan.ScanAll(&records, rows)
		if err != nil {
			return nil, err
		}
		_ = records
	}
	return records, nil
}
```