@georgysavva 2024-03-24
@zolstein here is the new release with the change: https://github.com/georgysavva/scany/releases/tag/v2.1.1